Завершил 15 задание с лонг

+++ Laba2--task15/Laba2--task15.cpp
@@ -10,7 +10,7 @@ using namespace std;
 
 int main() {
     setlocale(LC_ALL, "Rus");
-    double k;
+    long double k;
     cout << "Введите положительное количество грибов\n";
     cin >> k;
